+++ lib/src/models/utils/two_lists.dart
@@ -1,5 +1,5 @@
 class TwoLists<T> {
-  final List<T> listDups;
-  final List<T> listNoDups;
-  const TwoLists({required this.listDups, required this.listNoDups});
+  final List<T> dups;
+  final List<T> noDups;
+  const TwoLists({required this.dups, required this.noDups});
 }